Fix new view preprepare set (starting from 1) + fix JSON handling of new-view/view-change
